+++ src/button/demos/icon.vue
@@ -1,29 +1,22 @@
 <template>
-  <div>
-    <div class="tdesign-demo-block">
-      <t-button variant="text" :icon="starIcon">带图标文字按钮</t-button>
-      <t-button theme="primary">
-        <template #icon>
-          <t-icon-star />
-        </template>
-        带图标的按钮
-      </t-button>
-    </div>
-    <div class="tdesign-demo-block">
-      <t-button loading>带图标的按钮-加载</t-button>
-    </div>
+  <div class="flex align-center margin-right">
+    <t-button theme="primary" :icon="iconFunc" :disabled="disabled"></t-button>
+    <t-button theme="primary" shape="round" :icon="iconFunc" :disabled="disabled"></t-button>
+  </div>
+  <div class="flex align-center margin-right">
+    <t-button theme="primary" shape="square" :icon="iconFunc" :disabled="disabled"></t-button>
+    <t-button theme="primary" shape="circle" :icon="iconFunc" :disabled="disabled"></t-button>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue';
-import { StarIcon as TIconStar } from 'tdesign-icons-vue-next';
+<script lang="ts" setup>
+import { toRefs, defineProps, h } from 'vue';
+import { AppIcon as TIconApp } from 'tdesign-icons-vue-next';
 
-export default defineComponent({
-  components: { TIconStar },
-  setup() {
-    const starIcon = () => TIconStar;
-    return { starIcon };
-  },
+const props = defineProps({
+  disabled: Boolean,
 });
+
+const { disabled } = toRefs(props);
+const iconFunc = () => h(TIconApp);
 </script>