+++ src/button/demos/size.vue
@@ -1,17 +1,35 @@
 <template>
-  <div class="tdesign-demo-block">
-    <t-button size="large">大按钮</t-button>
+  <div class="flex align-center margin-right">
+    <t-button size="large" theme="primary">按钮 44</t-button>
+    <t-button theme="primary">按钮 40</t-button>
+    <t-button size="small" theme="primary">按钮 36</t-button>
   </div>
-  <div class="tdesign-demo-block">
-    <t-button theme="primary" size="medium">中按钮</t-button>
+  <div class="flex align-center margin-right">
+    <!-- <t-button size="large" :icon="iconFunc" shape="round" theme="primary">按钮</t-button> -->
+    <t-button size="large" theme="primary" shape="round">
+      <template #icon>
+        <t-icon-app />
+      </template>
+      按钮
+    </t-button>
+    <t-button :icon="iconFunc" shape="round" theme="primary">按钮</t-button>
+    <t-button size="small" :icon="iconFunc" shape="round" theme="primary">按钮</t-button>
   </div>
-  <div class="tdesign-demo-block">
-    <t-button theme="primary" size="small">小按钮</t-button>
+  <div class="flex align-center margin-right">
+    <t-button size="large" :icon="iconFunc" shape="square" theme="primary"></t-button>
+    <t-button :icon="iconFunc" shape="square" theme="primary"></t-button>
+    <t-button size="small" :icon="iconFunc" shape="square" theme="primary"></t-button>
+  </div>
+  <div class="flex align-center margin-right">
+    <t-button size="large" :icon="iconFunc" shape="circle" theme="primary"></t-button>
+    <t-button :icon="iconFunc" shape="circle" theme="primary"></t-button>
+    <t-button size="small" :icon="iconFunc" shape="circle" theme="primary"></t-button>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue';
+<script lang="ts" setup>
+import { h } from 'vue';
+import { AppIcon as TIconApp } from 'tdesign-icons-vue-next';
 
-export default defineComponent({});
+const iconFunc = () => h(TIconApp);
 </script>