+++ src/tag/demos/shape.vue
@@ -1,15 +1,4 @@
 <template>
-  <div class="tdesign-mobile-demo">
-    <div class="tag-demo">
-      <t-tag theme="primary">直角</t-tag>
-      <t-tag theme="primary" shape="round">圆角</t-tag>
-      <t-tag theme="primary" shape="mark">半圆</t-tag>
-    </div>
-  </div>
+  <t-tag theme="primary" shape="round">圆角</t-tag>
+  <t-tag theme="primary" shape="mark">半圆</t-tag>
 </template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-export default defineComponent({});
-</script>