+++ src/tag/demos/variant.vue
@@ -1,16 +1,6 @@
 <template>
-  <div class="tdesign-mobile-demo">
-    <div class="tag-demo">
-      <t-tag theme="primary" variant="dark">深色</t-tag>
-      <t-tag theme="primary" variant="light">浅色</t-tag>
-      <t-tag theme="primary" variant="outline">描边</t-tag>
-      <t-tag theme="primary" variant="light-outline">浅色描边</t-tag>
-    </div>
-  </div>
+  <t-tag theme="primary" variant="dark">深色</t-tag>
+  <t-tag theme="primary" variant="light">浅色</t-tag>
+  <t-tag theme="primary" variant="outline">描边</t-tag>
+  <t-tag theme="primary" variant="light-outline">浅色描边</t-tag>
 </template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-export default defineComponent({});
-</script>