+++ src/tag/tag.vue
@@ -1,18 +1,20 @@
 <template>
-  <span :class="classes" :style="style" :aria-disabled="disabled" role="button" @click="handleClick">
+  <span :class="classes" :style="tagStyle" :aria-disabled="disabled" role="button" @click="handleClick">
     <span :class="`${baseClass}__icon`">
       <t-node :content="iconContent"></t-node>
     </span>
     <span :class="`${baseClass}__text`">
       <t-node :content="tagContent"></t-node>
     </span>
-    <close-icon v-if="closable && !disabled" :class="`${baseClass}__close`" @click="onClickClose" />
+    <span v-if="closable && !disabled" :class="`${baseClass}__icon-close`" @click="onClickClose">
+      <close-icon />
+    </span>
   </span>
 </template>
 
 <script lang="ts">
 import { CloseIcon } from 'tdesign-icons-vue-next';
-import { defineComponent, computed, toRefs, getCurrentInstance, SetupContext } from 'vue';
+import { defineComponent, computed, getCurrentInstance, SetupContext } from 'vue';
 import config from '../config';
 import TagProps from './props';
 import { useEmitEvent, renderContent, renderTNode, TNode } from '../shared';
@@ -22,7 +24,10 @@ const name = `${prefix}-tag`;
 
 const Tag = defineComponent({
   name,
-  components: { CloseIcon, TNode },
+  components: {
+    CloseIcon,
+    TNode,
+  },
   props: TagProps,
   emits: ['close', 'click'],
   setup(props, context: SetupContext) {
@@ -31,44 +36,41 @@ const Tag = defineComponent({
     const tagContent = computed(() => renderContent(internalInstance, 'default', 'content'));
     const iconContent = computed(() => renderTNode(internalInstance, 'icon'));
     const baseClass = name;
-    const { size, shape, theme, variant, maxWidth, disabled, closable } = toRefs(props);
 
-    const style: { maxWidth?: string } = {};
-    if (maxWidth?.value) {
-      style.maxWidth = `${maxWidth.value}px`;
-    }
+    const tagStyle = computed(() => {
+      return props.maxWidth
+        ? { maxWidth: typeof props.maxWidth === 'number' ? `${props.maxWidth}px` : props.maxWidth }
+        : {};
+    });
 
     const classes = computed(() => [
       `${baseClass}`,
-      `${baseClass}-${theme.value}`,
-      `${baseClass}--${shape.value}`,
+      `${baseClass}--theme-${props.theme}`,
+      `${baseClass}--shape-${props.shape}`,
+      `${baseClass}--variant-${props.variant}`,
+      `${baseClass}--size-${props.size}`,
       {
-        [`${baseClass}-${variant.value}`]: variant.value,
-        [`${prefix}-is-error`]: theme.value === 'danger',
-        [`${prefix}-is-success`]: theme.value === 'success',
-        [`${prefix}-is-warning`]: theme.value === 'warning',
-        [`${prefix}-is-closable ${baseClass}--closable`]: closable.value,
-        [`${prefix}-is-disabled ${baseClass}--disabled`]: disabled.value,
-        [`${baseClass}--size-${size.value}`]: size.value,
+        [`${prefix}-is-closable ${baseClass}--closable`]: props.closable,
+        [`${prefix}-is-disabled ${baseClass}--disabled`]: props.disabled,
       },
     ]);
 
-    function onClickClose(e: Event) {
+    const onClickClose = (e: MouseEvent): void => {
       if (!props.disabled) {
-        emitEvent('close', e);
+        emitEvent('close', { e });
       }
-    }
+    };
 
     const handleClick = (e: MouseEvent): void => {
       if (!props.disabled) {
-        emitEvent('click', e);
+        emitEvent('click', { e });
       }
     };
 
     return {
       baseClass,
       classes,
-      style,
+      tagStyle,
       onClickClose,
       handleClick,
       iconContent,