+++ src/toast/demos/mobile.vue
@@ -21,21 +21,33 @@
     </tdesign-demo-block>
     <tdesign-demo-block title="02 展示位置和展示时间" summary="弹窗展示位置为顶部、中部、底部三种，展示时间可自定义">
       <div class="toast-demo">
-        <t-button block size="large" variant="outline" @click="showPosition('top')">顶部Top</t-button>
-        <t-button block size="large" variant="outline" @click="showPosition('middle')">中间Middle</t-button>
-        <t-button block size="large" variant="outline" @click="showPosition('bottom')">底部Bottom</t-button>
+        <t-button size="large" variant="outline" @click="showPosition('top')">顶部Top</t-button>
+        <t-button size="large" variant="outline" @click="showPosition('middle')">中间Middle</t-button>
+        <t-button size="large" variant="outline" @click="showPosition('bottom')">底部Bottom</t-button>
+        <t-button size="large" variant="outline" @click="showDuration(5000)">显示 5 秒 </t-button>
       </div>
     </tdesign-demo-block>
-    <tdesign-demo-block title="03 显示遮罩" summary="弹窗可显示遮罩，禁止滑动和点击">
+    <tdesign-demo-block title="03 显示遮罩" summary="弹窗可显示遮罩，可配置 Overlay 属性">
       <div class="toast-demo">
-        <t-button block size="large" variant="outline" @click="showOverlay">禁止滑动和点击</t-button>
+        <t-button size="large" variant="outline" @click="showOverlay">弹窗显示遮罩</t-button>
+      </div>
+    </tdesign-demo-block>
+
+    <tdesign-demo-block title="04 透传 Overlay" summary="向 Overlay 遮罩透传属性">
+      <div class="toast-demo">
+        <t-button size="large" variant="outline" @click="showOverlayProps">可滑动和点击透明遮罩</t-button>
+      </div>
+    </tdesign-demo-block>
+    <tdesign-demo-block title="05 禁止滑动和点击" summary="弹窗禁止滑动和点击">
+      <div class="toast-demo">
+        <t-button size="large" variant="outline" @click="showPreventScrollThrough">禁止滑动和点击</t-button>
       </div>
     </tdesign-demo-block>
   </div>
 </template>
 
 <script lang="ts">
-import { ref, h, defineComponent } from 'vue';
+import { h, defineComponent } from 'vue';
 import { UserIcon } from 'tdesign-icons-vue-next';
 import Toast from '../index';
 
@@ -108,8 +120,23 @@ export default defineComponent({
     showPosition(placement: TdToastProps['placement']) {
       Toast({ placement, message: '轻提示内容' });
     },
+    showDuration(duration: number) {
+      Toast({ message: `轻提示显示 ${duration}ms`, duration });
+    },
     showOverlay() {
+      Toast({ theme: 'loading', message: '加载中...', showOverlay: true });
+    },
+    showPreventScrollThrough() {
       Toast({ theme: 'loading', message: '加载中...', preventScrollThrough: true });
+    },
+    showOverlayProps() {
+      Toast({
+        theme: 'loading',
+        message: '加载中...',
+        showOverlay: true,
+        overlayProps: { transparent: true, visible: true },
+        preventScrollThrough: false,
+      });
     },
   },
 });