+++ src/toast/toast.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <t-overlay v-show="preventScrollThrough" />
+    <t-overlay v-bind="customOverlayProps" />
     <div :class="classes">
       <t-node :content="iconContent"></t-node>
       <div v-if="messageContent" :class="`${name}__text`">
@@ -12,7 +12,7 @@
 
 <script lang="ts">
 import { LoadingIcon, CheckCircleIcon, ErrorCircleIcon } from 'tdesign-icons-vue-next';
-import { computed, toRefs, ref, defineComponent, getCurrentInstance, h, onMounted, onUnmounted } from 'vue';
+import { computed, toRefs, ref, defineComponent, getCurrentInstance, h } from 'vue';
 import { renderTNode, TNode } from '../shared';
 import TOverlay from '../overlay';
 import ToastProps from './props';
@@ -53,20 +53,22 @@ export default defineComponent({
       },
     ]);
 
-    const cls = `${prefix}-overflow-hidden`;
-    onMounted(() => {
-      props.preventScrollThrough && document.body.classList.add(cls);
-    });
+    const toastOverlayProps = {
+      preventScrollThrough: props.preventScrollThrough,
+      visible: props.showOverlay,
+    };
 
-    onUnmounted(() => {
-      props.preventScrollThrough && document.body.classList.remove(cls);
-    });
+    const customOverlayProps = computed(() => ({
+      ...props.overlayProps,
+      ...toastOverlayProps,
+    }));
 
     return {
       name: ref(name),
       classes,
       iconContent,
       messageContent,
+      customOverlayProps,
       ...toRefs(props),
     };
   },